Add support for NpStr based file content and support for transaction based file access.
For transaction based file access the file content is the result of a
transaction, such as a field bus read transaction and possibly the new
file content is processed in a resulting transaction, such as a field
bus write access.

Corrections in formatting of the ASCII serialization of numbers.
Consequently now a terminating NL character is visible.